backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
ezdxf==1.3.4
Pillow==10.4.0
python-dotenv==1.0.1
